+++ llvm/docs/ReleaseNotes.rst
@@ -67,6 +67,11 @@ Changes to the LLVM IR
   type is now mandatory. Specify `i8* null` to migrate from the obsoleted
   2-field form.
 
+* The ``byval`` attribute can now take a type parameter:
+  ``byval(<ty>)``. If present it must be identical to the argument's
+  pointee type. In the next release we intend to make this parameter
+  mandatory in preparation for opaque pointer types.
+
 Changes to the ARM Backend
 --------------------------